Case: Removed vertical mount
Vertical mount is a point of failure for the acrylic (especially as
when mounted, the lid has huge leverage to snap the mount notch).  When
the screw holes are adjusted for Model B+ compatibility, this we be
even worse.  Additionally, it may have little practical use, as the
case can stand vertically.

May revisit at a later time.