acrylic - Merged lids (Gamer & Creator)
Previously, the gamer lid and creator lid were independent.  With the latest
design, the only difference between the two is that the Gamer requires holes
for mounting a board on the top, and requires clearance to feed the ribbon
cable through the lid.

This design widens the lid connector hole, and adds a grid of 1" spaced
mounting holes.